+++ hsql/src/main/java/org/testifyproject/resource/hsql/InMemoryHSQLResource.java
@@ -20,18 +20,18 @@ import java.sql.Connection;
 import java.sql.SQLException;
 import javax.sql.DataSource;
 import org.hsqldb.jdbc.JDBCDataSource;
+import org.testifyproject.LocalResourceProvider;
 import org.testifyproject.ResourceInstance;
-import org.testifyproject.ResourceProvider;
 import org.testifyproject.TestContext;
 import org.testifyproject.core.ResourceInstanceBuilder;
 
 /**
- * An implementation of ResourceProvider that provides an in-memory HSQL
+ * An implementation of LocalResourceProvider that provides an in-memory HSQL
  * DataSource and connection.
  *
  * @author saden
  */
-public class InMemoryHSQLResource implements ResourceProvider<JDBCDataSource, DataSource, Connection> {
+public class InMemoryHSQLResource implements LocalResourceProvider<JDBCDataSource, DataSource, Connection> {
 
     private JDBCDataSource server;
     private Connection client;
@@ -52,8 +52,8 @@ public class InMemoryHSQLResource implements ResourceProvider<JDBCDataSource, Da
             server = dataSource;
             client = dataSource.getConnection();
 
-            return new ResourceInstanceBuilder<DataSource, Connection>()
-                    .server(server, "inmemoryHSQLDataSource", DataSource.class)
+            return ResourceInstanceBuilder.builder()
+                    .instance(server, "inmemoryHSQLDataSource", DataSource.class)
                     .client(client, "inmemoryHSQLConnection", Connection.class)
                     .build();
         } catch (SQLException e) {